requirements.txt
torch
matplotlib
tqdm
numpy
sklearn
Pillow
pycocotools
